Get past history purchases
Returns the most recent purchases made by the user for each SKU, even if that purchase is expired, canceled, or consumed.

Because of the note from Android :
Note: The getPurchaseHistory() method has higher overhead than getPurchases(), because it requires a call to the Google Play server. You should use getPurchases() if you do not actually need the user's purchase history.

This request is made in an AsyncTask and IBillingPastHistoryHandler is updated when is done.
Also is not called in onServiceConnected, but only by user's request.

The user is loading the history with loadPastHistoryPurchasesFromGoogle :
public void  loadPastHistoryPurchasesFromGoogle(IBillingPastHistoryHandler callback) {
        new LoadPastHistoryTask(callback).execute();
    }

There is a new interface just for that call IBillingPastHistoryHandler :
public interface IBillingPastHistoryHandler {
        void onPastHistoryRestored();
        void onPastHistoryError();
    }

And when the call is done the user can get the history skus from :
public List<String> listPastHistoryProducts() {
        return cachedProductsPast.getContents();
    }

public List<String> listPastHistorySubscriptions() {
        return cachedSubscriptionsPast.getContents();
    }
And details from :
public TransactionDetails getPastPurchaseTransactionDetails(String productId) {
        return getPurchaseTransactionDetails(productId, cachedProductsPast);
    }

    public TransactionDetails getPastSubscriptionTransactionDetails(String productId) {
        return getPurchaseTransactionDetails(productId, cachedSubscriptionsPast);
    }

The minimum Google_API_Version that support this feature is 6.
There is a method to check that too :
public boolean isFullHistorySupported()

I have added in the sample app 1 button to preview the feature :
				bp.loadPastHistoryPurchasesFromGoogle(new BillingProcessor.IBillingPastHistoryHandler() {
                    @Override
                    public void onPastHistoryRestored() {
                        showToast("Past history loaded!");
                        for (String sku : bp.listPastHistoryProducts()) {
                            TransactionDetails td = bp.getPastPurchaseTransactionDetails(sku);
                            Log.d(LOG_TAG, "Past Managed Product: " + sku + " :\n" + td.toString());
                        }
                        for (String sku : bp.listPastHistorySubscriptions()) {
                            TransactionDetails td = bp.getPastSubscriptionTransactionDetails(sku);
                            Log.d(LOG_TAG, "Past Subscription: " + sku + " :\n" + td.toString());
                        }
                    }

                    @Override
                    public void onPastHistoryError() {
                        showToast("Failed to load past history");
                    }
                });